requirements.txt
python=3.10
langchain
langchain_core
langchain_groq
langchain_community
langchain_text_splitters
langchain_huggingface
langchain_classic
streamlit
python-dotenv 